feat(heureka): Refactor controllers and reduce redundencies
