+++ heureka/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "heureka",
-  "version": "2.3.1",
+  "version": "2.3.2",
   "author": "UI-Team",
   "contributors": [
     "Hoda Noori, Arturo Reuschenbach Pucernau"
